requirements.txt
crython==0.2.0
paho-mqtt==1.5.0
pycoingecko==1.3.0
six==1.13.0
